ch05_ai_agent/requirements.txt
strands-agents
strands-agents-tools
bedrock-agentcore
aws-opentelemetry-distro>=0.10.0
boto3
opentelemetry-instrumentation-logging
PyYAML
